+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Differences to the Plugins WordPress version
 
-* none
+* new option in Search Control: collapsed
+* filter render_block for core/shortcode is no longer needed
 
 Please note the [install instructions](https://leafext.de/en/doku/about/versions/).